refactor: extract class RequiredVersion to its own domain; update test patch paths
