+++ dbt_platform_helper/utils/versioning.py
@@ -3,8 +3,6 @@ import subprocess
 from pathlib import Path
 
 from dbt_platform_helper.constants import DEFAULT_TERRAFORM_PLATFORM_MODULES_VERSION
-from dbt_platform_helper.platform_exception import PlatformException
-from dbt_platform_helper.providers.io import ClickIOProvider
 from dbt_platform_helper.providers.platform_helper_versioning import (
     PlatformHelperVersioning,
 )
@@ -14,67 +12,6 @@ from dbt_platform_helper.providers.semantic_version import VersionStatus
 from dbt_platform_helper.providers.validation import ValidationException
 from dbt_platform_helper.providers.version import GithubVersionProvider
 from dbt_platform_helper.providers.yaml_file import YamlFileProvider
-
-
-# TODO to be moved into domain
-class PlatformHelperVersionNotFoundException(PlatformException):
-    def __init__(self):
-        super().__init__(f"""Platform helper version could not be resolved.""")
-
-
-# TODO to be moved into domain
-class RequiredVersion:
-    def __init__(self, io=None, platform_helper_versioning=None):
-        self.io = io or ClickIOProvider()
-        self.platform_helper_versioning = platform_helper_versioning or PlatformHelperVersioning(
-            io=self.io
-        )
-
-    def get_required_platform_helper_version(
-        self, pipeline: str = None, version_status: PlatformHelperVersionStatus = None
-    ) -> str:
-        pipeline_version = version_status.pipeline_overrides.get(pipeline)
-        version_precedence = [
-            pipeline_version,
-            version_status.platform_config_default,
-            version_status.deprecated_version_file,
-        ]
-        non_null_version_precedence = [
-            f"{v}" if isinstance(v, SemanticVersion) else v for v in version_precedence if v
-        ]
-
-        out = non_null_version_precedence[0] if non_null_version_precedence else None
-
-        if not out:
-            raise PlatformHelperVersionNotFoundException
-
-        return out
-
-    def get_required_version(self, pipeline=None):
-        version_status = self.platform_helper_versioning.get_status()
-        self.io.process_messages(version_status.validate())
-        required_version = self.get_required_platform_helper_version(pipeline, version_status)
-        self.io.info(required_version)
-        return required_version
-
-    # Used in the generate command
-    def check_platform_helper_version_mismatch(self):
-        if not running_as_installed_package():
-            return
-
-        version_status = self.platform_helper_versioning.get_status()
-        self.io.process_messages(version_status.validate())
-
-        required_version = SemanticVersion.from_string(
-            self.get_required_platform_helper_version(version_status=version_status)
-        )
-
-        if not version_status.local == required_version:
-            message = (
-                f"WARNING: You are running platform-helper v{version_status.local} against "
-                f"v{required_version} specified for the project."
-            )
-            self.io.warn(message)
 
 
 # TODO to be removed after config tests are updated - temporary wrapper mid-refactor
